+++ src/main/java/com/prenotapp/_repo/ICategoryRepo.java
@@ -2,4 +2,4 @@ package com.prenotapp._repo;
 
 import com.prenotapp._model.Category;
 
-public interface ICategoryRepo extends IGenericRepo<Category, Integer> {}
+public interface ICategoryRepo extends IGenericRepo<Category, Long> {}
