+++ src/main/java/com/prenotapp/_service/ICategoryService.java
@@ -2,4 +2,4 @@ package com.prenotapp._service;
 
 import com.prenotapp._model.Category;
 
-public interface ICategoryService extends ICRUD<Category, Integer> {}
+public interface ICategoryService extends ICRUD<Category, Long> {}
