+++ src/main/java/com/prenotapp/_service/IPersonaService.java
@@ -2,4 +2,4 @@ package com.prenotapp._service;
 
 import com.prenotapp._model.Persona;
 
-public interface IPersonaService extends ICRUD<Persona, Integer> {}
+public interface IPersonaService extends ICRUD<Persona, Long> {}
